refactor(app): move wagmi=config out of wallet/evm/config.svelte.ts
