+++ app2/src/lib/wallet/evm/config.svelte.ts
@@ -1,154 +1,26 @@
 import {
   connect as _connect,
   type Connector,
-  createConfig,
-  createStorage as createWagmiStorage,
-  deserialize,
   disconnect as _disconnect,
-  fallback,
   getAccount,
   type GetAccountReturnType,
-  http,
   reconnect,
-  serialize,
   switchChain as _switchChain,
-  unstable_connector,
   watchAccount
 } from "@wagmi/core"
-import { coinbaseWallet, injected, metaMask, walletConnect } from "@wagmi/connectors"
-import {
-  arbitrumSepolia,
-  berachainTestnetbArtio,
-  holesky,
-  scrollSepolia,
-  sepolia
-} from "@wagmi/core/chains"
 import { AddressEvmCanonical } from "$lib/schema/address"
 import { wallets } from "$lib/stores/wallets.svelte"
 import { Effect, Option } from "effect"
 import { TESTNET_APP_INFO } from "$lib/config/app"
 import type { Hex } from "viem"
 
-export const chains = [
-  sepolia,
-  holesky,
-  berachainTestnetbArtio,
-  arbitrumSepolia,
-  scrollSepolia
-] as const
-export type ConfiguredChainId = (typeof chains)[number]["id"]
+import { wagmiConfig, type ConfiguredChainId } from "./wagmi-config"
+import { sepolia } from "@wagmi/core/chains"
 
 export type Wallet = GetAccountReturnType
 export type ConnectorType = "injected" | "walletConnect"
 
-const WALLETCONNECT_PROJECT_ID = "49fe74ca5ded7142adefc69a7788d14a"
 const LAST_CONNECTED_WALLET_KEY = "last-connected-wallet"
-
-export const configSvelte = createConfig({
-  chains,
-  cacheTime: 4_000,
-  pollingInterval: 4_000,
-  syncConnectedChain: true,
-  batch: { multicall: true },
-  multiInjectedProviderDiscovery: true,
-  transports: {
-    [sepolia.id]: fallback([
-      unstable_connector(injected, {
-        retryCount: 3,
-        retryDelay: 100,
-        key: "unstable_connector-injected-sepolia",
-        name: "unstable_connector-injected-sepolia"
-      }),
-      http(`https://rpc.11155111.sepolia.chain.kitchen`, {
-        name: "Chain Kitchen - Sepolia"
-      }),
-      http(sepolia.rpcUrls.default.http.at(0), { name: "default Sepolia RPC" })
-    ]),
-    [holesky.id]: fallback([
-      unstable_connector(injected, {
-        retryCount: 3,
-        retryDelay: 100,
-        key: "unstable_connector-injected-holesky",
-        name: "unstable_connector-injected-holesky"
-      }),
-      http(`https://rpc.17000.holesky.chain.kitchen`, {
-        name: "Chain Kitchen - Holesky"
-      }),
-      http(holesky.rpcUrls.default.http.at(0), { name: "default Holesky RPC" })
-    ]),
-    [berachainTestnetbArtio.id]: fallback([
-      unstable_connector(injected, {
-        retryCount: 3,
-        retryDelay: 100,
-        key: "unstable_connector-injected-berachain",
-        name: "unstable_connector-injected-berachain"
-      }),
-      http(berachainTestnetbArtio.rpcUrls.default.http.at(0), { name: "default Berachain RPC" })
-    ]),
-    [arbitrumSepolia.id]: fallback([
-      unstable_connector(injected, {
-        retryCount: 3,
-        retryDelay: 100,
-        key: "unstable_connector-injected-arbitrum-sepolia",
-        name: "unstable_connector-injected-arbitrum-sepolia"
-      }),
-      http(arbitrumSepolia.rpcUrls.default.http.at(0), { name: "default Arbitrum Sepolia RPC" })
-    ]),
-    [scrollSepolia.id]: fallback([
-      unstable_connector(injected, {
-        retryCount: 3,
-        retryDelay: 100,
-        key: "unstable_connector-injected-scroll-sepolia",
-        name: "unstable_connector-injected-scroll-sepolia"
-      }),
-      http(scrollSepolia.rpcUrls.default.http.at(0), { name: "default Scroll Sepolia RPC" })
-    ])
-  },
-  storage: createWagmiStorage({
-    serialize,
-    deserialize,
-    key: "union-wagmi",
-    storage: typeof window !== "undefined" ? window.localStorage : undefined
-  }),
-  connectors: [
-    injected({
-      shimDisconnect: true,
-      unstable_shimAsyncInject: 2_500
-    }),
-    coinbaseWallet({
-      darkMode: true,
-      appName: TESTNET_APP_INFO.name,
-      appLogoUrl: TESTNET_APP_INFO.iconUrl,
-      enableMobileWalletLink: true
-    }),
-    metaMask({
-      injectProvider: true,
-      dappMetadata: {
-        name: TESTNET_APP_INFO.name,
-        url: TESTNET_APP_INFO.baseUrl,
-        iconUrl: TESTNET_APP_INFO.iconUrl
-      },
-      useDeeplink: true
-    }),
-    walletConnect({
-      projectId: WALLETCONNECT_PROJECT_ID,
-      showQrModal: true,
-      metadata: {
-        name: TESTNET_APP_INFO.name,
-        description: "Connect via WalletConnect",
-        url: TESTNET_APP_INFO.baseUrl,
-        icons: [TESTNET_APP_INFO.iconUrl]
-      }
-    })
-  ]
-})
-
-configSvelte.subscribe(
-  state => state.chainId,
-  _chainId => {
-    console.log("config state changed", _chainId)
-  }
-)
 
 function getLastConnectedWalletId(): string | undefined {
   if (typeof window !== "undefined") {
@@ -188,7 +60,7 @@ class SepoliaStore {
       this.unwatchAccount = undefined
     }
 
-    this.unwatchAccount = watchAccount(configSvelte, {
+    this.unwatchAccount = watchAccount(wagmiConfig, {
       onChange: account => {
         if (this.connectionStatus === "connected" && !account.address && this.address) {
           return
@@ -213,7 +85,7 @@ class SepoliaStore {
       }
 
       const result = await evmConnect(walletId, sepolia.id)
-      const account = getAccount(configSvelte)
+      const account = getAccount(wagmiConfig)
       setLastConnectedWalletId(account.connector?.id)
 
       this.updateAccount({
@@ -265,16 +137,16 @@ class SepoliaStore {
     try {
       this.connectionStatus = "connecting"
 
-      const lastConnector = configSvelte.connectors.find(c => c.id === lastWalletId)
+      const lastConnector = wagmiConfig.connectors.find(c => c.id === lastWalletId)
       if (!lastConnector) {
         console.warn("Last connected connector not found:", lastWalletId)
         this.connectionStatus = "disconnected"
         return
       }
 
-      await reconnect(configSvelte, { connectors: [lastConnector] })
+      await reconnect(wagmiConfig, { connectors: [lastConnector] })
 
-      const account = getAccount(configSvelte)
+      const account = getAccount(wagmiConfig)
       this.updateAccount({
         chain: account.chain?.name ?? "sepolia",
         address: account.address,
@@ -316,7 +188,7 @@ export const sepoliaStore = new SepoliaStore()
 
 const excludeWalletList = ["io.leapwallet.LeapWallet", "app.keplr"]
 
-export const evmWalletsInformation = configSvelte.connectors
+export const evmWalletsInformation = wagmiConfig.connectors
   .map(connector => {
     const id = connector.id.toLowerCase()
     const name = connector.name.toLowerCase()
@@ -344,20 +216,20 @@ export async function evmConnect(
   evmWalletId: EvmWalletId,
   chainId: ConfiguredChainId = sepolia.id
 ) {
-  const connector = configSvelte.connectors.find(connector => connector.id === evmWalletId)
+  const connector = wagmiConfig.connectors.find(connector => connector.id === evmWalletId)
   if (connector) {
-    return await _connect(configSvelte, { connector, chainId })
+    return await _connect(wagmiConfig, { connector, chainId })
   }
   throw new Error(`Connector ${evmWalletId} not found`)
 }
 
 export async function evmDisconnect() {
   try {
-    const connector = getAccount(configSvelte).connector
+    const connector = getAccount(wagmiConfig).connector
     if (connector) {
-      await _disconnect(configSvelte, { connector })
+      await _disconnect(wagmiConfig, { connector })
     } else {
-      await _disconnect(configSvelte)
+      await _disconnect(wagmiConfig)
     }
   } catch (error) {
     console.error("Error during disconnect:", error)
@@ -365,5 +237,4 @@ export async function evmDisconnect() {
   }
 }
 
-export const evmSwitchChain = (chainId: ConfiguredChainId) =>
-  _switchChain(configSvelte, { chainId })
+export const evmSwitchChain = (chainId: ConfiguredChainId) => _switchChain(wagmiConfig, { chainId })
