+++ app2/src/lib/wallet/evm/index.ts
@@ -1,7 +1,6 @@
 export {
-  chains,
   evmConnect,
   sepoliaStore,
   type EvmWalletId,
   evmWalletsInformation
-} from "./config.svelte.ts"
+} from "./config.svelte"
